requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch==2.2.2
numpy==1.26.4
pyenchant==3.2.2
python-multipart==0.0.9